Reorganized pubnub client file structure
